Added timebound mode
Used multithreading for the timer and taking input simultaneously. This update took 3 days because I even tried to make the speech into a different thread when the game mode starts so the "Guess higher" or "Guess lower" prompts don't waste the player's time. But pyttsx3 was really cranky with this. Then I tried using gTTS and pygame for this, but for pygame when I imported it in timebound module and tried to suppress the intro speech of pygame--which worked but this threw an error in main.py as I used os.devnull. So I finally decided to not give a damn and just
add speech prompts in between the game-- this does waste (very little) time of the player, but it works. In future I will fix this too when I upgrade this project w UI, or start using Godot.