Added Gaslight mode
Gaslight mode is similar to normal mode, but the twist is that the game randomly lies to you twice per game if the number is 'higher' or 'lower'-- and you'll never know when it's lying and when it's telling the truth. The code logic wasn't difficult, I just used random.sample(range(1, 7), 2) so the game randomly lies from guess to 7, and does this twice.