Finished TPG 4.0
Added music for endgame mode, one for the intro where Kate acts sane, other for the entire game mode.

Fixed bugs where a lot of files weren't being opened if the main.py was opened independently instead of opening the entire game folder. Fixed this using the os module.

Reset all the .txt files which saved the game data as well as all the highscore files, so this version is now fully playable-- you can download the files and modules and directly start playing

This commit marks the end of TPG 4.0.
Now I'll start working on the last update for this game ; TPG 5.0 will be the final polish to this game adding UI and pixel characters to the game.