Added dialogue system and difficulty selection screen
The dialogue system added to normal game mode will be used in all other modes as well.

This dialogue system is dynamic; helps to render the text according to the font size and the dialogue box size so the text doesn't go out of the dialogue box .

Added difficulty selection screen for normal game mode and used the dialogue system there. The code for this screen will be used for all game modes here on.

Also edited and removed most of the original code (from TPG 4.0) because I would need to revamp almost all of it due to pygame. This revamped version of normal game mode will be the base code for all other game modes in the future.