First playable version of TPG 5.0 normal game mode
This commit marks the first ever playable beta version of The Perfect Guess.

Basic game logic of TPG 4.0 was converted to UI with a much cleaner and efficient codes following DRY principle.

There's still a lot of polish to be added to this game mode like adding a post-game scoreboard, mysterious voice hints every 4 games, a live update on top of screen for guess no. of the player, highscore update after the game ends. But this is relatively easy to implement. Once the normal game mode is fully polished, it'll act as the base code for all other game modes, making their development process a whole lot easier.