Finished timebound mode
I'm proud to announce that this commit marks the first and final, fully polished and playable timebound mode.

The mode features difficulty options (range of numbers) as well as an option for player to select the timer between 20, 30, 40 secs.

Added a timer beep sound when timer reaches zero, which wasn't present in TPG 4.0.

Made the timer display in Aseprite and added/implemented it in the game along with the buttons for 20/30/40secs.

Edited the buttons for Yes/No because the buttons prior to editing didn't share the same font size as the other buttons did.

In this mode, just like TPG 4.0 I've used a thread which counts the time and shares it with the main thread using a dictionary. I've raised an exception when the timer runs out so matter what the player is doing or if the game is in an inner loop, the game will still end.