Minor feature update for timebound mode
Prior to this update the timebound mode gave the player option to select the timer from 20/30/40 secs.

This commit adds an option for 10secs too, for the competitive players out there. This option is available for all difficulty levels as well.

Speaking on the technical side, this feature addition was pretty easy, I just needed to make a button for 10secs in Aseprite. After that I needed to add the png to a button class and change the position of other option's buttons to adjust the 10sec button on the screen, and lastly update Kate's prompt for selecting between the options.