Minor audio update for Timebound mode mode
This commit adds a ticking sound for each second in the timebound mode. Prior to this, only the bg music played during the game and a time out beep when the timer went to zero along with a losing sound effect. But now the ticking sound effects each second adds tension.

Had to set the volume for sound effect to 0.3 and edit the audio module to set the music volume of timebound mode to 1.2 as they both were clashing each other which caused the music to set on a very low volume.

Minor bug fix:
If the timer ran out, only Kate's image blitted and she said her prompt, because of this the earlier blitted image and current used to get merge and it looked like Kate had 3 hands.
Fixed this by blitting the bg again before blitting Kate's img after the timer runs out.