Finished Gaslight mode
This commit marks the first and final, fully playable and polished Gaslight mode. In this mode Kate randomly lies twice if the number is higher/lower. To do this I used random.sample(range(n1, n2), 2). I varied n1, n2 according to the difficulty  selected by the player.

This mode works fully synchronous with endgame mode as a high score (varies for each difficulty) in any difficulty level unlocks Endgame Mode. And this is locked after the player plays Endgame mode as it's supposed to be a one-time event.

Made a new sound effect, added and implemented it for when Endgame Mode unlocks.

Lowered the volume for glitch sound effect (played when Endgame isn't unlocked and player clicks on "???" button) to 0.1 as it sounded too sharp and annoying when system sound was above 40%.