Menu update- TPG 5.1 : The Aftermath Update
This is the initial commit for the 5.1 Aftermath update for TPG. This update will change the game permanently and add a permadeath to Kate if the player fails to save her in Endgame mode.

The menu screen was updated such that if the player loses in Endgame (which quits the game) and re-opens the game, the aftermath screen will appear. But I needed to make sure that happens only if the player loses, and that too once. To keep track of whether player has won/lost/not yet played Endgame we already have endgame_result.txt, but to keep track of whether Aftermath screen was played, I needed to make an aftermath_counter.txt. Now the aftermath screen in menu only occurs if endgame_result.txt == "lost" and aftermath_counter.txt == "" and then inside this if block itself we open aftermath_counter.txt and write "lock" in it so that the screen is never played again. The logic for the aftermath screen is pretty simple, I fill the screen with black, then zoom in "The Aftermath" text using a for loop and play a cinematic thud sound at the beginning of this.

Removed the tts() and only_tts() functions from menu as they were carried on from TPG 4.0's file and served no purpose in menu after the 5.0 UI update.