Timebound mode update- TPG 5.1 : The Aftermath Update
This commit marks a fully playable version for Timebound mode with the Aftermath update integrated within it. Which means if player fails to save Kate in Endgame then she's nowhere to be seen in this mode.

The logic to this stays the same as used in Normal mode for the Aftermath update.