Gaslight mode update- TPG 5.1 : The Aftermath Update
This commit marks the final update for the 5.1 update of TPG.

Gaslight mode is now fully playable along with the Aftermath update integrated within it. Which means if player fails to save Kate in Endgame then she's nowhere to be seen in the entire game now, including Gaslight mode.

The logic to make this work remains the same as used for normal/timebound mode.