Made and added icon for the game
The game no more uses the default pygame icon and instead uses my own custom made icon from Aseprite, which is a png.

I have also added an .ico file of the same icon which will be used later when I package the game into an exe so that the icon also appears in the taskbar.