Normal mode update- TPG 5.2 : The Customs Update
This commit marks the first update for TPG 5.2 : The Customs Update.
With this commit, normal mode now supports a full custom difficulty in which player can choose the range from which they want to guess the number from instead of guessing it from the default easy/mid/hard modes.

Custom mode doesn't have a highscore system unlike other pre-defined modes, for obvious reasons.

Post-game scoreboard is different for custom difficulty. In pre-defined difficulty it showed current score and highscore. In custom difficulty it shows custom range that was choses along with attempts taken (current score).

Custom mode comes with its own set of roasts and prompts.
This update is fully polished and edge-case ready for scenarios where : player enters a negative number, player enters the larger number in range first and the smaller number afterwards, player enters the same numbers in range, etc along with custom randomized roast sets for each.

The prompt additions made in Custom mode obviously come with 5.1 : Aftermath Update included in them, which means these prompts won't be seen if Kate dies in Endgame.

File changes :

1. Made a CUSTOM button in Aseprite, added/implemented it.

2. Made a custom_input() function in normal_mode.py which is used to take input of both numbers in range and will be used in the other timebound/gaslight mode too when they're integrated with this update.
Highscore system is now locked behind an if difficulty != "custom" block to prevent errors.

2. common_prompts.py was edited to add more functions which give different randomized prompts/roasts for custom mode as well as its edge cases.

Logic :
The custom_input() function returns a number if it's not negative or an alphabet. So I run the function one time to first when CUSTOM button is clicked, after that when I get the first number, then run the same function again to get a second number. After that our n stays the same (so I didn't have to edit anything from the actual game loop besides post game prompts) because I simply do n = random.randint(n1, n2) and then end the difficulty selection loop. But here this random.randint is in a try block (for edge case, if n1>n2). Now I could've just added an if n1<n2 : n = random.randint(n1, n2) else : n = random.randint(n2, n1) and called it a day. But this game's trademark is how it handles edge cases and roasts players for each type of edge case. So I put the randint function in a try block and made a except ValueError block which called customized and randomized roast for this edge case (this happens if endgame_result != "lost", but if endgame_result == "lost" then there's just one Jokingo prompt which tells the player to enter the numbers in increasing order).