Minor packaging update
The exe file was throwing error after quitting the game, hence had to change quit() --> sys.exit() in all files.

Made it easier to unlock Endgame mode by increasing the highscore attempt by 3 for each gaslight difficulty. Did this because I though Endgame is the pinnacle as well as the ending to my game, so I have to ensure that all players unlock it instead of gate-keeping it.