Minor difficulty changes
I earlier changed made it easier to unlock Endgame. But I realised that it's still relatively easy only on easy difficulty, but on mid/hard difficulty it doesn't make any sense. Therefore I edited the value for mid (1-500) from 11 to 15 and hard (1-1000) from 13 to 20 attempts in order to unlock Endgame. For easy (1-100) the attempts to unlock still stays 9. Earlier I thought that it would be easy to unlock Endgame even at 6,8,10 attempts for each difficulty resp. But I was wrong, after playtesting the game I realised it's extremely difficulty to get those scores especially when Kate is lying to you twice, and you don't even know when she is. Even after these difficulty changes made, it's still difficult to unlock Endgame, but it's do-able in a few hours.