Difficulty changes for Endgame mode
Despite cutting the attempts from 25 to 15 and truth serums from 7 to 5, Endgame still felt too easy.

I've now cut down the truth serums from 5 to 3. Now it truly feels difficult to save Kate, which is exactly what I was going for. It's still not impossible but you need a lot of skills AND luck to win Endgame now.